Added INJECTATRON.
SPIR-V byte-code will be injected into the resulting binary rather than
loaded from disk. This reduces the dependency on the file-system which
does not play particularly well with Bazel.

The INJECTATRON is a set of rules and a simple binary that generates a
C++ header file that defines a std::vector<char> of the specified
file's contents. This can be used in place of loading a file from disk
into a std::vector<char>.